Preview and Send Email Using Mailable Classes

+++ app/Http/Controllers/JobController.php
@@ -3,7 +3,11 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use App\Models\User;
 use App\Models\Job;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -28,16 +32,22 @@ class JobController extends Controller
             'salary' => ['required']
         ]);
 
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
 
+        Mail::to($job->employer->user->email)->send(
+            new \App\Mail\JobPosted($job)
+        );
+
+
         return redirect('/jobs');
     }
 
     public function edit(Job $job) {
+
         return view('jobs.edit', ['job' => $job]);
     }
 
@@ -56,8 +66,10 @@ class JobController extends Controller
     }
 
     public function destroy(Job $job) {
-        $job->delete();
 
+        Gate::authorize('edit', $job);
+
+        $job->delete();
 
      return redirect('/jobs');
     }
